+++ tests/api/test_hmm.py
@@ -67,7 +67,7 @@ class TestFind:
 
         hmm_document["hidden"] = False
 
-        await client.db.hmm.insert(hmm_document)
+        await client.db.hmm.insert_one(hmm_document)
 
         resp = await client.get("/api/hmms")
