+++ virtool/api/groups.py
@@ -113,9 +113,9 @@ async def remove(req):
 
     group_id = req.match_info["group_id"]
 
-    document = await db.groups.find_one_and_delete({"_id": group_id})
+    delete_result = await db.groups.delete_one({"_id": group_id})
 
-    if not document:
+    if not delete_result.deleted_count:
         return not_found()
 
     await virtool.db.groups.update_member_users(db, group_id, remove=True)
